// src/service/daftarSekolah/dto.ts
export default interface DaftarSekolahAttributes {
    id: string;
    category: string;
    name: string;
    address: string;
    postal_code: string;
    provinsi_id: number;
    kabupaten_id: number;
    phone_number: string;
    email: string;
    facebook: string;
    total_students: number;
}